Save pending callbacks for each outgoing query and immediately call all the pending callbacks for a connection if an 'error' event is emitted on a Thrift connection.
Previously we didn't map those errors back (we just logged them) which means
that if there was an outgoing query it wouldn't fail before query timeout is
reached (defaults to 5 seconds).

With this change, the callback is called as soon as the Thrift error is emitted
which means that when using PooledConnection fail over is
basically instantaneous.